chore: add merchant location model
