+++ internal/utils/common.go
@@ -20,3 +20,17 @@ func ToInt(ni sql.NullString) int64 {
 	}
 	return 0
 }
+
+func String2Int64(input string, defaultValue int64) int64 {
+	if parsed, err := strconv.ParseInt(input, 10, 64); err == nil {
+		return parsed
+	}
+	return defaultValue
+}
+
+func String2Int32(input string, defaultValue int) int {
+	if parsed, err := strconv.ParseInt(input, 10, 32); err == nil {
+		return int(parsed)
+	}
+	return defaultValue
+}
